#72 make different join, if no geometry is involved
The query has to be generated differently, if the search should filter
for a geometry. In this case also the tables flurstueck and geom has to
be joined, as a geometry has to be fetched.
Although if no geometry is involved at all, then this join leads to a
wrong result, as some Buchungsblaetter do not have a reference to a
flurstueck. In this case the tables flurstueck and geom must not be
joined.